Move block sync extension related files to separated folder
